# Force color in tests for `PkgEval`