- Fix https://github.com/JuliaPlots/UnicodePlots.jl/issues/134.

```bash
$ julia --color=yes -e 'using Pkg; pkg"test UnicodePlots"'
$ julia --color=no  -e 'using Pkg; pkg"test UnicodePlots"'
```
, until https://github.com/KristofferC/Crayons.jl/issues/47 is fixed.

- Code cleanup, use `NamedTuple` instead of `Dict` (faster):

**before**
```julia
julia> using UnicodePlots
julia> @time lineplot([0, 4, 8], [10, 1, 10], color = :blue, name = "other line")
  1.096257 seconds (858.07 k allocations: 52.733 MiB, 4.19% gc time, 99.74% compilation time)
julia> @time lineplot([0, 4, 8], [10, 1, 10], color = :blue, name = "other line")
  0.000232 seconds (1.09 k allocations: 29.000 KiB)
```

**after**
```julia
julia> using UnicodePlots
@time lineplot([0, 4, 8], [10, 1, 10], color = :blue, name = "other line")
  0.922900 seconds (857.27 k allocations: 52.692 MiB, 1.76% gc time, 99.84% compilation time)
@time lineplot([0, 4, 8], [10, 1, 10], color = :blue, name = "other line")
  0.000172 seconds (1.09 k allocations: 29.000 KiB)
```